test(cloud-client): assert https_only by behaviour, and pin only what Debug renders
Copilot's review was right: reqwest 0.13.4's Client Debug renders neither
connect_timeout nor https_only, so listing them on the reference builder pinned
nothing while the assertion message claimed the whole helper was verified.

Probed what Debug actually renders. Only two of the four settings appear:
redirect_policy: "Policy(Custom)" and TotalTimeout. So:

- The reference builder now carries only the timeout, and the doc comment and
  assertion messages say exactly what the comparison covers.
- https_only, the one with a security consequence, is asserted by behaviour
  instead: both constructors' clients must refuse a plain-http origin that is
  answering 200, and the stub must record no request. Removing https_only from
  build_http_client fails that test.
- connect_timeout is not observable either way; the doc comment says so rather
  than implying it is covered.

The stub server moves to a shared test_support module so the construction tests
can use it too.
