+++ crates/spice-cloud-client/src/lib.rs
@@ -25,6 +25,8 @@ mod client;
 pub mod endpoints;
 pub mod error;
 pub mod redirect;
+#[cfg(test)]
+mod test_support;
 pub mod types;
 
 pub use client::CloudClient;
