+++ crates/spice-cloud-client/src/redirect.rs
@@ -82,11 +82,8 @@ pub fn same_origin_redirect_policy() -> reqwest::redirect::Policy {
 #[cfg(test)]
 mod tests {
     use super::{MAX_REDIRECTS, is_same_origin, same_origin_redirect_policy};
+    use crate::test_support::{Stub, ok_with, redirect_to};
     use reqwest::{StatusCode, Url};
-    use std::io::{Read, Write};
-    use std::net::{SocketAddr, TcpListener, TcpStream};
-    use std::sync::{Arc, Mutex};
-    use std::thread::JoinHandle;
 
     fn url(value: &str) -> Url {
         Url::parse(value).expect("test URL should parse")
@@ -158,123 +155,6 @@ mod tests {
     /// `reqwest` sanitises on a cross-origin hop, so nothing but the policy keeps it on
     /// the origin it was minted for.
     const API_KEY: &str = "test-api-key-value";
-
-    /// A one-connection-at-a-time HTTP/1.1 stub, enough to answer a redirect script.
-    ///
-    /// The workspace `tokio` carries no `net` feature, so this is a blocking `std::net`
-    /// listener on its own thread. Every response closes the connection, so each hop
-    /// arrives as its own accept and the recorded order is the request order.
-    struct Stub {
-        addr: SocketAddr,
-        requests: Arc<Mutex<Vec<String>>>,
-        worker: Option<JoinHandle<()>>,
-    }
-
-    impl Stub {
-        /// Answer each request with `respond(nth)`, where `nth` is 1-based, until dropped.
-        fn serve(respond: impl Fn(usize) -> String + Send + 'static) -> Self {
-            let listener =
-                TcpListener::bind("127.0.0.1:0").expect("stub should bind a loopback port");
-            let addr = listener
-                .local_addr()
-                .expect("stub should report its address");
-            let requests = Arc::new(Mutex::new(Vec::new()));
-
-            let recorded = Arc::clone(&requests);
-            let worker = std::thread::spawn(move || {
-                for stream in listener.incoming() {
-                    let Ok(mut stream) = stream else { break };
-                    let head = read_request_head(&mut stream);
-                    // The drop poke connects and sends nothing; that is the stop signal.
-                    if head.is_empty() {
-                        break;
-                    }
-
-                    let nth = {
-                        let mut recorded =
-                            recorded.lock().expect("request log should not be poisoned");
-                        recorded.push(head);
-                        recorded.len()
-                    };
-
-                    let _ = stream.write_all(respond(nth).as_bytes());
-                    let _ = stream.flush();
-                }
-            });
-
-            Self {
-                addr,
-                requests,
-                worker: Some(worker),
-            }
-        }
-
-        fn url(&self, path: &str) -> String {
-            let addr = self.addr;
-            format!("http://{addr}{path}")
-        }
-
-        /// The request heads seen so far, in order, lower-cased so an assertion does not
-        /// depend on how the client happens to case a header name on the wire.
-        fn requests(&self) -> Vec<String> {
-            self.requests
-                .lock()
-                .expect("request log should not be poisoned")
-                .clone()
-        }
-    }
-
-    impl Drop for Stub {
-        fn drop(&mut self) {
-            // Unblock the accept the worker is parked in, then let it finish.
-            let _ = TcpStream::connect(self.addr);
-            if let Some(worker) = self.worker.take() {
-                let _ = worker.join();
-            }
-        }
-    }
-
-    /// Read one request's head, stopping at the blank line that ends it. These requests
-    /// carry no body, so nothing after it needs consuming.
-    fn read_request_head(stream: &mut TcpStream) -> String {
-        /// Far above any head these tests produce, so reaching it means the peer is not
-        /// sending one and the read should stop rather than accumulate without a bound.
-        const MAX_HEAD_BYTES: usize = 16 * 1024;
-
-        let mut head = Vec::new();
-        let mut chunk = [0_u8; 256];
-
-        loop {
-            match stream.read(&mut chunk) {
-                Ok(0) | Err(_) => break,
-                Ok(read) => {
-                    head.extend_from_slice(&chunk[..read]);
-                    if head.len() >= MAX_HEAD_BYTES
-                        || head.windows(4).any(|window| window == b"\r\n\r\n")
-                    {
-                        break;
-                    }
-                }
-            }
-        }
-
-        String::from_utf8_lossy(&head).to_lowercase()
-    }
-
-    fn redirect_to(location: &str) -> String {
-        format!(
-            "HTTP/1.1 307 Temporary Redirect\r\nLocation: {location}\r\n\
-             Content-Length: 0\r\nConnection: close\r\n\r\n"
-        )
-    }
-
-    fn ok_with(body: &str) -> String {
-        let length = body.len();
-        format!(
-            "HTTP/1.1 200 OK\r\nContent-Length: {length}\r\n\
-             Connection: close\r\n\r\n{body}"
-        )
-    }
 
     /// A client that differs from the default in nothing but the policy under test and a
     /// deadline.
